Support for ILI9486 with ArduinoGFX and better implementation seperation
For ILI9341, we should be able to use ILI9341_t3 for all microcontrollers (even non-teensy), not sure about ILI9488_t3.

Arduino_GFX supports a lot of screens and microcontrollers, so it should be possible to use this library as a fallback/catchall. Right now it's used to support ILI9486.

Removed the IDENTITY_R_* macros, as they weren't needed in the end

Lots of refactoring to IdentityImplementation.h so that all the different screen libraries are supported in a cleaner and more flexible way.